1R-micro-python/requirements.txt
Flask
pytest
PyYAML
requests